[fix] #94 검증 중복 검증 관ë련 불필요한 로직 및 import문 삭제
