+++ module-domain/domain-inventories/src/main/java/com/chaing/domain/inventories/service/inbound/FranchiseInboundService.java
@@ -5,7 +5,7 @@ import com.chaing.domain.inventories.dto.command.FranchiseInboundCreateCommand;
 import com.chaing.domain.inventories.dto.raw.FranchiseInventoryRawData;
 import com.chaing.domain.inventories.usecase.inbound.executor.InboundExecutor;
 import com.chaing.domain.inventories.usecase.inbound.reader.InboundReader;
-import com.chaing.domain.inventories.usecase.inbound.valiator.InboundValidator;
+import com.chaing.domain.inventories.usecase.inbound.validator.InboundValidator;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Service;
