+++ module-domain/domain-inventories/src/main/java/com/chaing/domain/inventories/usecase/inbound/validator/FactoryInboundValidatorImpl.java
@@ -1,8 +1,8 @@
-package com.chaing.domain.inventories.usecase.inbound.valiator;
+package com.chaing.domain.inventories.usecase.inbound.validator;
 
 import com.chaing.core.enums.LogType;
-import com.chaing.domain.inventories.dto.command.FranchiseInboundCreateCommand;
-import com.chaing.domain.inventories.dto.raw.FranchiseInventoryRawData;
+import com.chaing.domain.inventories.dto.command.FactoryInboundCreateCommand;
+import com.chaing.domain.inventories.dto.raw.FactoryInventoryRawData;
 import com.chaing.domain.inventories.exception.InventoriesErrorCode;
 import com.chaing.domain.inventories.exception.InventoriesException;
 import com.chaing.domain.inventories.usecase.inbound.reader.InboundReader;
@@ -14,12 +14,12 @@ import java.time.LocalDate;
 import java.util.List;
 
 @Component
-@Qualifier("franchise")
+@Qualifier("factory")
 @RequiredArgsConstructor
-public class FranchiseInboundValidatorImpl implements InboundValidator<FranchiseInboundCreateCommand, FranchiseInventoryRawData> {
+public class FactoryInboundValidatorImpl implements InboundValidator<FactoryInboundCreateCommand, FactoryInventoryRawData> {
 
-    @Qualifier("franchise")
-    private final InboundReader<FranchiseInventoryRawData> inboundReader;
+    @Qualifier("factory")
+    private final InboundReader<FactoryInventoryRawData> inboundReader;
 
     public final int SERIAL_CODE_LENGTH = 10;
 
@@ -34,9 +34,20 @@ public class FranchiseInboundValidatorImpl implements InboundValidator<Franchise
     }
 
     @Override
-    public void checkScanValidity(FranchiseInboundCreateCommand command) {
+    public void checkScanValidity(FactoryInboundCreateCommand command) {
 
+        String serialCode = command.serialCode();
         LocalDate date = command.manufactureDate();
+
+        // 식별 코드 누락 확인
+        if(serialCode == null || serialCode.isBlank()) {
+            throw new InventoriesException(InventoriesErrorCode.INVENTORIES_SERIAL_CODE_IS_NULL);
+        }
+
+        // 식별 코드 유효성 확인
+        if(serialCode.length() != SERIAL_CODE_LENGTH) {
+            throw new InventoriesException(InventoriesErrorCode.INVALID_SERIAL_CODE);
+        }
 
         // 날짜 누락 확인
         if(date == null) {
@@ -47,37 +58,15 @@ public class FranchiseInboundValidatorImpl implements InboundValidator<Franchise
         if(date.isAfter(LocalDate.now())){
             throw new InventoriesException(InventoriesErrorCode.INVALID_MANUFACTURED_DATE);
         }
-
-        List<String> serialCodes = command.serialCodes();
-
-        // 리스트 누락 확인
-        if(serialCodes == null || serialCodes.isEmpty()) {
-            throw new InventoriesException(InventoriesErrorCode.INVENTORIES_SERIAL_CODE_IS_NULL);
-        }
-
-        serialCodes.forEach(serialCode -> {
-            // 식별 코드 누락 확인
-            if (serialCode == null || serialCode.isBlank()) {
-                throw new InventoriesException(InventoriesErrorCode.INVENTORIES_SERIAL_CODE_IS_NULL);
-            }
-
-            // 식별 코드 유효성 확인
-            if (serialCode.length() != SERIAL_CODE_LENGTH) {
-                throw new InventoriesException(InventoriesErrorCode.INVALID_SERIAL_CODE);
-            }
-        });
     }
 
     @Override
-    public void checkPendingDataExistence(List<FranchiseInventoryRawData> entities) {
+    public void checkPendingDataExistence(List<FactoryInventoryRawData> entities) {
         if(entities.isEmpty() || entities == null) {
             throw new InventoriesException(InventoriesErrorCode.INVENTORIES_IS_NULL);
         }
 
-        for (FranchiseInventoryRawData entity : entities) {
-            if(entity.franchiseId() == null) {
-                throw new InventoriesException(InventoriesErrorCode.INVENTORIES_IS_INVALID);
-            }
+        for (FactoryInventoryRawData entity : entities) {
             if(entity.status() == null) {
                 throw new InventoriesException(InventoriesErrorCode.INVENTORIES_IS_INVALID);
             }
@@ -85,9 +74,6 @@ public class FranchiseInboundValidatorImpl implements InboundValidator<Franchise
                 throw new InventoriesException(InventoriesErrorCode.INVENTORIES_IS_INVALID);
             }
             if(entity.manufactureDate() == null) {
-                throw new InventoriesException(InventoriesErrorCode.INVENTORIES_IS_INVALID);
-            }
-            if(entity.boxCode() == null) {
                 throw new InventoriesException(InventoriesErrorCode.INVENTORIES_IS_INVALID);
             }
         }
